Added map marker grouping
