+++ packages/website/package.json
@@ -44,6 +44,7 @@
     "firebase-tools": "^8.4.3",
     "immutable": "^4.0.0-rc.12",
     "leaflet": "^1.6.0",
+    "leaflet.markercluster": "^1.4.1",
     "lodash": "^4.17.15",
     "material-table": "^1.65.0",
     "moment": "^2.27.0",
@@ -53,6 +54,7 @@
     "react-dom": "^16.13.1",
     "react-hook-form": "^6.0.8",
     "react-leaflet": "^2.7.0",
+    "react-leaflet-markercluster": "^2.0.0",
     "react-player": "^2.3.0",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
