+++ packages/website/src/custom.d.ts
@@ -9,3 +9,5 @@ declare module "*.css" {
   const classNames: IClassNames;
   export = classNames;
 }
+
+declare module "react-leaflet-markercluster";
