+++ packages/website/src/routes/Homepage/Map/Markers/ColoredBuoyIcon.tsx
@@ -1,0 +1,8 @@
+import L from "leaflet";
+import {
+  colorFinder,
+  degreeHeatingWeeksCalculator,
+} from "../../../../helpers/degreeHeatingWeeks";
+import buoySvg from "./buoy.svg";
+
+// export default ColoredBuoyIcon;
